revise

+++ _9Method/ProblemSolvingCommandLineArguements/SumOfNumbersInCommandLine.java
@@ -5,15 +5,16 @@ public class SumOfNumbersInCommandLine {
     //so we pass the numbers that also consider as stirng so we have converet as double
     //There is a class called Double and have method "parse.Double"
     public static void main(String args[]){
-        int s=0;
+        double s=0;
         //use for loop use for understand the concept of for each loop
         for(String x:args){
             //if we pass string with number it throws error called number format error occur 
             //to overcome this we condition below
-            if(x.matches("[0-9\\.]+")){        s+=Double.parseDouble(x);
+            if(x.matches("[0-9\\.]+")){       
+                 s+=Double.parseDouble(x);
 
             }
         }
-        System.out.println("Sum is " +s);
+        System.out.println("Sum is "+s);
     }
 }
